+++ scriptworker/context.py
@@ -7,6 +7,7 @@ object, let's point to them from a single context object.
 
 Attributes:
     log (logging.Logger): the log object for the module.
+
 """
 import arrow
 from copy import deepcopy
@@ -31,8 +32,6 @@ class Context(object):
             FrozenDict.
         credentials_timestamp (int): the unix timestamp when we last updated
             our credentials.
-        poll_task_urls (dict): contains the Azure ``queues`` urls and an ``expires``
-            datestring.
         proc (asyncio.subprocess.Process): when launching the script, this is
             the process object.
         queue (taskcluster.async.Queue): the taskcluster Queue object
@@ -41,11 +40,11 @@ class Context(object):
         task (dict): the task definition for the current task.
         temp_queue (taskcluster.async.Queue): the taskcluster Queue object
             containing the task-specific temporary credentials.
+
     """
 
     config = None
     credentials_timestamp = None
-    poll_task_urls = None
     proc = None
     queue = None
     session = None
@@ -66,6 +65,7 @@ class Context(object):
         When setting ``claim_task``, we also set ``self.task`` and
         ``self.temp_credentials``, zero out ``self.reclaim_task`` and ``self.proc``,
         then write a task.json to disk.
+
         """
         return self._claim_task
 
@@ -91,6 +91,7 @@ class Context(object):
 
         When setting credentials, also create a new ``self.queue`` and
         update self.credentials_timestamp.
+
         """
         if self._credentials:
             return dict(deepcopy(self._credentials))
@@ -106,6 +107,7 @@ class Context(object):
 
         Args:
             credentials (dict): taskcluster credentials.
+
         """
         if credentials:
             return Queue({
@@ -123,6 +125,7 @@ class Context(object):
         ``reclaim_task`` will be ``None`` if there hasn't been a claimed task yet,
         or if a task has been claimed more recently than the most recent
         reclaimTask call.
+
         """
         return self._reclaim_task
 
@@ -137,6 +140,7 @@ class Context(object):
         """dict: The latest temp credentials, or None if we haven't claimed a task yet.
 
         When setting, create ``self.temp_queue`` from the temp taskcluster creds.
+
         """
         if self._temp_credentials:
             return dict(deepcopy(self._temp_credentials))
@@ -153,6 +157,7 @@ class Context(object):
             path (str): the path to write to
             contents (dict): the contents of the json blob
             message (str): the message to log
+
         """
         log.debug(message.format(path=path))
         makedirs(os.path.dirname(path))