+++ scriptworker/exceptions.py
@@ -10,6 +10,7 @@ class ScriptWorkerException(Exception):
 
     Attributes:
         exit_code (int): this is set to 5 (internal-error).
+
     """
 
     exit_code = 5
@@ -20,6 +21,7 @@ class ScriptWorkerGPGException(ScriptWorkerException):
 
     Attributes:
         exit_code (int): this is set to 5 (internal-error).
+
     """
 
     exit_code = 5
@@ -30,6 +32,7 @@ class ScriptWorkerRetryException(ScriptWorkerException):
 
     Attributes:
         exit_code (int): this is set to 4 (resource-unavailable)
+
     """
 
     exit_code = 4
@@ -50,6 +53,7 @@ class ScriptWorkerTaskException(ScriptWorkerException):
 
     Attributes:
         exit_code (int): this is 1 by default (failure)
+
     """
 
     def __init__(self, *args, exit_code=1, **kwargs):
@@ -60,6 +64,7 @@ class ScriptWorkerTaskException(ScriptWorkerException):
             exit_code (int, optional): The exit_code we should exit with when
                 this exception is raised.  Defaults to 1 (failure).
             **kwargs: These are passed on via super().
+
         """
         self.exit_code = exit_code
         super(ScriptWorkerTaskException, self).__init__(*args, **kwargs)
@@ -70,6 +75,7 @@ class DownloadError(ScriptWorkerTaskException):
 
     Attributes:
         exit_code (int): this is set to 4 (resource-unavailable).
+
     """
 
     def __init__(self, msg):
@@ -77,6 +83,7 @@ class DownloadError(ScriptWorkerTaskException):
 
         Args:
             msg (string): the error message
+
         """
         super(DownloadError, self).__init__(
             msg, exit_code=4
@@ -88,6 +95,7 @@ class CoTError(ScriptWorkerTaskException, KeyError):
 
     Attributes:
         exit_code (int): this is set to 3 (malformed-payload).
+
     """
 
     def __init__(self, msg):
@@ -95,6 +103,7 @@ class CoTError(ScriptWorkerTaskException, KeyError):
 
         Args:
             msg (string): the error message
+
         """
         super(CoTError, self).__init__(
             msg, exit_code=3