+++ scriptworker/test/test_cot_verify.py
@@ -44,7 +44,7 @@ def chain(rw_context):
     }
     rw_context.task = {
         'scopes': ['project:releng:signing:cert:nightly-signing', 'ignoreme'],
-        'dependencies': [],
+        'dependencies': ['decision_task_id'],
         'provisionerId': rw_context.config['provisioner_id'],
         'schedulerId': 'schedulerId',
         'workerType': rw_context.config['worker_type'],
@@ -77,7 +77,7 @@ def build_link(chain):
         'provisionerId': 'provisioner',
         'workerType': 'workerType',
         'scopes': [],
-        'dependencies': [],
+        'dependencies': ['some_task_id'],
         'metadata': {
             'source': 'https://hg.mozilla.org/mozilla-central',
         },
@@ -257,7 +257,13 @@ def test_raise_on_errors(errors, raises):
     {'payload': {}, 'provisionerId': 'test-dummy-provisioner', 'workerType': 'test-dummy-myname', 'scopes': ["x"]},
     'scriptworker', False
 ), (
+    {'payload': {'mounts': [], 'osGroups': []}, 'provisionerId': '', 'workerType': '', 'scopes': []},
+    'generic-worker', False
+), (
     {'payload': {'image': 'x'}, 'provisionerId': 'test-dummy-provisioner', 'workerType': '', 'scopes': []},
+    None, True
+), (
+    {'payload': {'image': 'x', 'osGroups': []}, 'provisionerId': '', 'workerType': '', 'scopes': []},
     None, True
 )))
 def test_guess_worker_impl(chain, task, expected, raises):
@@ -630,14 +636,20 @@ def test_verify_link_in_task_graph(chain, decision_link, build_link):
     cotverify.verify_link_in_task_graph(chain, decision_link, build_link)
 
 
-def test_verify_link_in_task_graph_fuzzy_match(chain, decision_link, build_link):
+@pytest.mark.parametrize("zero_deps", (True, False))
+def test_verify_link_in_task_graph_fuzzy_match(chain, decision_link, build_link, zero_deps):
     chain.links = [decision_link, build_link]
+    task_defn1 = deepcopy(build_link.task)
+    task_defn2 = deepcopy(chain.task)
+    if zero_deps:
+        build_link.task['dependencies'] = []
+        chain.task['dependencies'] = []
     decision_link.task_graph = {
         'bogus-task-id': {
-            'task': deepcopy(build_link.task)
+            'task': task_defn1
         },
         'bogus-task-id2': {
-            'task': deepcopy(chain.task)
+            'task': task_defn2
         }
     }
     cotverify.verify_link_in_task_graph(chain, decision_link, build_link)
@@ -862,6 +874,12 @@ async def test_verify_docker_worker_task(mocker):
     mocker.patch.object(cotverify, 'check_interactive_docker_worker', new=noop_sync)
     mocker.patch.object(cotverify, 'verify_docker_image_sha', new=noop_sync)
     await cotverify.verify_docker_worker_task(mock.MagicMock(), mock.MagicMock())
+
+
+# verify_generic_worker_task {{{1
+@pytest.mark.asyncio
+async def test_verify_generic_worker_task(mocker):
+    await cotverify.verify_generic_worker_task(mock.MagicMock(), mock.MagicMock())
 
 
 # verify_worker_impls {{{1