+++ scriptworker/worker.py
@@ -3,14 +3,15 @@
 
 Attributes:
     log (logging.Logger): the log object for the module.
+
 """
 import aiohttp
+import arrow
 import asyncio
 import logging
 import os
 import sys
 
-from scriptworker.poll import find_task, get_azure_urls, update_poll_task_urls
 from scriptworker.artifacts import upload_artifacts
 from scriptworker.config import get_context_from_cmdln
 from scriptworker.constants import STATUSES
@@ -18,8 +19,8 @@ from scriptworker.cot.generate import generate_cot
 from scriptworker.cot.verify import ChainOfTrust, verify_chain_of_trust
 from scriptworker.gpg import get_tmp_base_gpg_home_dir, is_lockfile_present, rm_lockfile
 from scriptworker.exceptions import ScriptWorkerException
-from scriptworker.task import complete_task, reclaim_task, run_task, worst_level
-from scriptworker.utils import cleanup, retry_request, rm
+from scriptworker.task import claim_work, complete_task, reclaim_task, run_task, worst_level
+from scriptworker.utils import cleanup, rm
 
 log = logging.getLogger(__name__)
 
@@ -35,23 +36,20 @@ async def run_loop(context, creds_key="credentials"):
 
     Returns:
         int: status
+        None: if no task run.
+
     """
     loop = asyncio.get_event_loop()
-    await update_poll_task_urls(
-        context, context.queue.pollTaskUrls,
-        args=(context.config['provisioner_id'], context.config['worker_type']),
-    )
-    for poll_url, delete_url in get_azure_urls(context):
-        try:
-            claim_task_defn = await find_task(context, poll_url, delete_url,
-                                              retry_request)
-        except ScriptWorkerException:
-            await asyncio.sleep(context.config['poll_interval'])
-            break
-        if claim_task_defn:
+    tasks = await claim_work(context)
+    status = None
+    if tasks:
+        # Assume only a single task, but should more than one fall through,
+        # run them sequentially.  A side effect is our return status will
+        # be the status of the final task run.
+        for task_defn in tasks.get('tasks', []):
+            context.claim_task = task_defn
             log.info("Going to run task!")
             status = 0
-            context.claim_task = claim_task_defn
             loop.create_task(reclaim_task(context, context.task))
             try:
                 if context.config['verify_chain_of_trust']:
@@ -73,13 +71,14 @@ async def run_loop(context, creds_key="credentials"):
             await complete_task(context, status)
             cleanup(context)
             await asyncio.sleep(1)
-            return status
-    else:
-        await asyncio.sleep(context.config['poll_interval'])
+    await asyncio.sleep(context.config['poll_interval'])
+    return status
 
 
 async def async_main(context):
-    """Main async loop, following http://docs.taskcluster.net/queue/worker-interaction/ .
+    """Run the main async loop.
+
+    http://docs.taskcluster.net/queue/worker-interaction/
 
     This is a simple loop, mainly to keep each function more testable.
 
@@ -101,6 +100,7 @@ async def async_main(context):
 def main():
     """Scriptworker entry point: get everything set up, then enter the main loop."""
     context, credentials = get_context_from_cmdln(sys.argv[1:])
+    log.info("Scriptworker starting up at {} UTC".format(arrow.utcnow().format()))
     cleanup(context)
     conn = aiohttp.TCPConnector(limit=context.config['aiohttp_max_connections'])
     loop = asyncio.get_event_loop()